240201 3 #10845; queue made of vector. pop function by iterator not by erase.
O(n) : time complexity of erase(0) for queue pop. It cause to move all the elements.
O(1) : time complexity of front iterator +1 for queue pop. popped element doesn't erased so memory waste exists. But there is no element duplication, so it is faster than erase.  We sacrifice the memory for the speed.

BUT There are limitation that if the element of vector is changed, iterator variable doesn't work so we have to always reallocate iterator when elements are changed.
So I build this by another integer variable i, and always re-define iterator pointing front as queue.begin() + i when I manipulate the vector elements. pop causes i++.