240211 8 #1463 Solved; first DP. whenever dpTable updated, recursion must be done.
I build the code to do recursion when dpTable is initialized by first.
But it is a wrong implementation. Whenever dpTable is updated, we need to recall itself.
When value is updated as new efficient way exists, Re-calculating have to be done in the problem.
Only recalling when data is inserted in dpTable at first, causes that we could refer outdated dpTable so it is impossible to get the optimized answer.
I wasted my time because, although the entire dpTable is right when I review consequence output of code, the answer was wrong.
It happens because It refers outdated data when working; that can be seen by output.