240807 1 #14919 solved; Decimal
deal with the decimal by using string.
Error occurred SO many times, even if it seems no logical false in the code.
It was because of the location of `!cin.eof()`. the while loop was not terminated at the right time. it was the reason of the error, not the calculating logic itself of this.