240811 2 #1916  solved; Dijkstra
Dijkstra by adj matrix: O(N^2), with a help of G.

In Dijkstra, the next node should be the node with the minimum total cost in the current state, not the minimum cost edge.
Latter doesn't select the shortest path.  Dijkstra should select the shortest path based on the 'current state', in this code, it is cost[i], not the graph.

Using priority Queue is more efficient: O(N longN)