240814 2 #1504 solved; Dijkstra
if the graph is straightly connected, and we should visit far from the end, we always have to visit there first no matter how it is far, as we should visit there and come back from there. (e.g. : target1 - start - target2 - end)
Because of this case, comparing the distance to target[2] is not a complete comparison.
Thus, comparison has to be done on the basis of total distace of 2 cases : visiting target1 for the first, or 2 for the first.