240908 1 #2467 solved; Binary Search
One is selected sequentially, and the other is selected by Binary search.
Error happened because initial value for minDiff was too small before. forgot to multiply 2 after rewriting.

Instead of the Binary Seach, 2-Pointer can be used also; one starting from the max, and the other starting from the min, moving one by one based on a sum of both.