240914 1 #20040 1CC; Union-Find
Union Find implemented by two vectors, 'rank' and 'parent'.

'parent' is used for the faster comparison. Elements in the same tree(group) have a same parent. Every elem should have its most higher parent as parent[elem]. Basically, default parent of elem is itself.

'rank' is used for the performance of Union-Find. It is related with the rank of tree. For performance, the tree that has smaller rank must be placed under what has bigger rank. When both of them have same rank, merge them, and add 1 to the above one; the rank of tree is increased.