README edited : about git commit
Information about git is added. Git pane is broken now. I need to use terminal for the git.

git config is reloaded everytime I login, so reset the config everytime before the commit if I want the commits are shown in the Github.

This is all for to show commits in the Github.